feature/telegram_user_has_password
The original code work well for Telegram-Users which do not use a password with their account. The proposed changes allow to store in /home/pi/password the Telegram-User password. The setup page still needs to be updated to support entering the password plus the code still needs to be thoroughly tested (I only tested it with some POC implementation).